Suppress console.warn output
We have no meaningful way of handling warnings anyway.  I've briefly
considered forwarding them to Emacs and showing them there, but it
doesn't seem like something users should be concerned with. The
warning mentioned in the ticket is the first one I've come across so
far and it is not a useful piece of information in our use case.

Closes #64