Fix performance issue with cursor translation
As per https://github.com/prettier/prettier/issues/4801 ,
formatWithCursor can be extremely slow and consume large amounts of
memory. We're actually doing something very similar to the algorithm
described in that ticket already, namely character-wise diffs. This
means that the cursor is actually fine if we just let Emacs manage the
position.