Ignore stray stdout and stderr output
Print a beacon string at the beginning of the response and skip to it
when reading the response.

This avoids communication with the remote process getting stuck in
cases where calling `prettier.format` emits anything to stdout/stderr,
as is the case with @prettier/plugin-ruby at least up to version
1.5.2.

Also, handle remote process timeouts better.