Enable json5 and json-stringify parsers by default
These two parsers were accidentally omitted from the default list.  In
particular, this broke the choice of parser for package.json files.
So add them to the default list.

This solves the original motivating factor for #92, but not #92
itself which is a request for enhanced debugging capabilities.

Original commit by @aspiers, re-committed here in an attempt to fix a
mysterious CI failure.